Make the order of the output deterministic
Instead of printing the dependencies in the order we get them from the
cargo metadata, we sort them by their package names.

This is helpful in situations where you use `cargo-license` to generate
a `NOTICES` file that gets checked into the repository and want to
enfore in CI that the file is up to date.

Since the previous output was in a seemingly random order and the
performance penalty for sorting is rather small, I decided to not hide
this behavior behind a flag.