Add -o/--output option for file output
Currently, there's no way to write program output to a file other than
using shell redirection. In some cases, such as containerized CI jobs
that may not have a shell, this leaves no way to write the output to file.

Adding a -o option allows users to write the output to a file without
a shell.